requirements.txt
django
python-dotenv
google-genai
google-generativeai
django-allauth
pillow
"django-allauth[socialaccount]"
dj-rest-auth